[back/refactor/count관련-int-to-long] reactionCount, commentCount, replyCount 타입 변경

See merge request s10-webmobile1-sub2/S10P12B110!86

+++ backend/src/main/java/com/ssafy/forest/domain/dto/response/ArticleResDto.java
@@ -21,10 +21,10 @@ public class ArticleResDto {
     private LocalDateTime createdAt;
     private LocalDateTime modifiedAt;
     private List<String> imageList;
-    private int reactionCount;
-    private int commentCount;
+    private long reactionCount;
+    private long commentCount;
 
-    public static ArticleResDto of(Article article,int commentCount, int reactionCount) {
+    public static ArticleResDto of(Article article,long commentCount, long reactionCount) {
         return new ArticleResDto(
             article.getId(),
             article.getMember().getId(),
